Move event loop infrastructure to common code
This makes code of EventLoop (including its task queue impl) shared
between JVM and Native making support and further fixes easier.
It it not actually used in JS and DCE should remove it.
