+++ kotlinx-coroutines-core/common/src/internal/LockFreeTaskQueue.kt
@@ -1,13 +1,12 @@
 /*
- * Copyright 2016-2018 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
+ * Copyright 2016-2019 JetBrains s.r.o. Use of this source code is governed by the Apache 2.0 license.
  */
 
 package kotlinx.coroutines.internal
 
 import kotlinx.atomicfu.*
 import kotlinx.coroutines.*
-import java.util.*
-import java.util.concurrent.atomic.*
+import kotlin.jvm.*
 
 private typealias Core<E> = LockFreeTaskQueueCore<E>
 
@@ -84,7 +83,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
     private val mask = capacity - 1
     private val _next = atomic<Core<E>?>(null)
     private val _state = atomic(0L)
-    private val array = AtomicReferenceArray<Any?>(capacity)
+    private val array = atomicArrayOfNulls<Any?>(capacity)
 
     init {
         check(mask <= MAX_CAPACITY_MASK)
@@ -115,7 +114,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
                 if ((tail + 2) and mask == head and mask) return ADD_FROZEN // overfull, so do freeze & copy
                 // If queue is Multi-Consumer then the consumer could still have not cleared element
                 // despite the above check for one free slot.
-                if (!singleConsumer && array[tail and mask] != null) {
+                if (!singleConsumer && array[tail and mask].value != null) {
                     // There are two options in this situation
                     // 1. Spin-wait until consumer clears the slot
                     // 2. Freeze & resize to avoid spinning
@@ -130,7 +129,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
                 val newTail = (tail + 1) and MAX_CAPACITY_MASK
                 if (_state.compareAndSet(state, state.updateTail(newTail))) {
                     // successfully added
-                    array[tail and mask] = element
+                    array[tail and mask].value = element
                     // could have been frozen & copied before this item was set -- correct it by filling placeholder
                     var cur = this
                     while(true) {
@@ -144,7 +143,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
     }
 
     private fun fillPlaceholder(index: Int, element: E): Core<E>? {
-        val old = array.get(index and mask)
+        val old = array[index and mask].value
         /*
          * addLast actions:
          * 1) Commit tail slot
@@ -156,7 +155,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
          * perform *unique* check that current placeholder is our to avoid overwriting another producer placeholder
          */
         if (old is Placeholder && old.index == index) {
-            array.set(index and mask, element)
+            array[index and mask].value = element
             // we've corrected missing element, should check if that propagated to further copies, just in case
             return this
         }
@@ -173,7 +172,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
             if (state and FROZEN_MASK != 0L) return REMOVE_FROZEN // frozen -- cannot modify
             state.withState { head, tail ->
                 if ((tail and mask) == (head and mask)) return null // empty
-                val element = array[head and mask]
+                val element = array[head and mask].value
                 if (element == null) {
                     // If queue is Single-Consumer, then element == null only when add has not finished yet
                     if (singleConsumer) return null // consider it not added yet
@@ -190,7 +189,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
                 if (_state.compareAndSet(state, state.updateHead(newHead))) {
                     // Array could have been copied by another thread and it is perfectly fine, since only elements
                     // between head and tail were copied and there are no extra steps we should take here
-                    array[head and mask] = null // now can safely put null (state was updated)
+                    array[head and mask].value = null // now can safely put null (state was updated)
                     return element // successfully removed in fast-path
                 }
                 // Multi-Consumer queue must retry this loop on CAS failure (another consumer might have removed element)
@@ -214,7 +213,7 @@ internal class LockFreeTaskQueueCore<E : Any>(
                     return next() // continue to correct head in next
                 }
                 if (_state.compareAndSet(state, state.updateHead(newHead))) {
-                    array[head and mask] = null // now can safely put null (state was updated)
+                    array[head and mask].value = null // now can safely put null (state was updated)
                     return null
                 }
             }
@@ -242,7 +241,8 @@ internal class LockFreeTaskQueueCore<E : Any>(
             var index = head
             while (index and mask != tail and mask) {
                 // replace nulls with placeholders on copy
-                next.array[index and next.mask] = array[index and mask] ?: Placeholder(index)
+                val value = array[index and mask].value ?: Placeholder(index)
+                next.array[index and next.mask].value = value
                 index++
             }
             next._state.value = state wo FROZEN_MASK
@@ -252,12 +252,12 @@ internal class LockFreeTaskQueueCore<E : Any>(
 
     // Used for validation in tests only
     fun <R> map(transform: (E) -> R): List<R> {
-        val res = ArrayList<R>(array.length())
+        val res = ArrayList<R>(capacity)
         _state.value.withState { head, tail ->
             var index = head
             while (index and mask != tail and mask) {
                 // replace nulls with placeholders on copy
-                val element = array[index and mask]
+                val element = array[index and mask].value
                 @Suppress("UNCHECKED_CAST")
                 if (element != null && element !is Placeholder) res.add(transform(element as E))
                 index++
