+++ kotlinx-coroutines-core/js/src/EventLoop.kt
@@ -8,8 +8,20 @@ import kotlin.coroutines.*
 
 internal actual fun createEventLoop(): EventLoop = UnconfinedEventLoop()
 
+internal actual fun nanoTime(): Long = unsupported()
+
 internal class UnconfinedEventLoop : EventLoop() {
-    override fun dispatch(context: CoroutineContext, block: Runnable) {
-        throw UnsupportedOperationException("runBlocking event loop is not supported")
-    }
+    override fun dispatch(context: CoroutineContext, block: Runnable): Unit = unsupported()
 }
+
+internal actual abstract class EventLoopImplPlatform : EventLoop() {
+    protected actual fun unpark(): Unit = unsupported()
+}
+
+internal actual object DefaultExecutor {
+    public actual fun enqueue(task: Runnable): Unit = unsupported()
+    public actual fun schedule(delayedTask: EventLoopImplBase.DelayedTask): Unit = unsupported()
+}
+
+private fun unsupported(): Nothing =
+    throw UnsupportedOperationException("runBlocking event loop is not supported")
