+++ kotlinx-coroutines-core/jvm/src/internal/ThreadSafeHeap.kt
@@ -4,156 +4,5 @@
 
 package kotlinx.coroutines.internal
 
-import kotlinx.coroutines.*
-import java.util.*
-
-/**
- * @suppress **This an internal API and should not be used from general code.**
- */
-@InternalCoroutinesApi
-public interface ThreadSafeHeapNode {
-    public var heap: ThreadSafeHeap<*>?
-    public var index: Int
-}
-
-/**
- * Synchronized binary heap.
- * @suppress **This an internal API and should not be used from general code.**
- */
-@InternalCoroutinesApi
-public class ThreadSafeHeap<T> : SynchronizedObject() where T: ThreadSafeHeapNode, T: Comparable<T> {
-    private var a: Array<T?>? = null
-
-    @JvmField @Volatile
-    public var size = 0
-
-    public val isEmpty: Boolean get() = size == 0
-
-    @Synchronized
-    public fun clear() {
-        Arrays.fill(a, 0, size, null)
-        size = 0
-    }
-
-    @Synchronized
-    public fun peek(): T? = firstImpl()
-
-    @Synchronized
-    public fun removeFirstOrNull(): T? =
-        if (size > 0) {
-            removeAtImpl(0)
-        } else {
-            null
-        }
-
-    // @Synchronized // NOTE! NOTE! NOTE! inline fun cannot be @Synchronized
-    public inline fun removeFirstIf(predicate: (T) -> Boolean): T? = synchronized(this) {
-        val first = firstImpl() ?: return null
-        if (predicate(first)) {
-            removeAtImpl(0)
-        } else {
-            null
-        }
-    }
-
-    @Synchronized
-    public fun addLast(node: T) = addImpl(node)
-
-    // @Synchronized // NOTE! NOTE! NOTE! inline fun cannot be @Synchronized
-    public inline fun addLastIf(node: T, cond: () -> Boolean): Boolean = synchronized(this) {
-        if (cond()) {
-            addImpl(node)
-            true
-        } else {
-            false
-        }
-    }
-
-    @Synchronized
-    public fun remove(node: T): Boolean {
-        return if (node.heap == null) {
-            false
-        } else {
-            val index = node.index
-            assert { index >= 0 }
-            removeAtImpl(index)
-            true
-        }
-    }
-
-    @PublishedApi
-    internal fun firstImpl(): T? = a?.get(0)
-
-    @PublishedApi
-    internal fun removeAtImpl(index: Int): T {
-        assert { size > 0 }
-        val a = this.a!!
-        size--
-        if (index < size) {
-            swap(index, size)
-            val j = (index - 1) / 2
-            if (index > 0 && a[index]!! < a[j]!!) {
-                swap(index, j)
-                siftUpFrom(j)
-            } else {
-                siftDownFrom(index)
-            }
-        }
-        val result = a[size]!!
-        assert { result.heap === this }
-        result.heap = null
-        result.index = -1
-        a[size] = null
-        return result
-    }
-
-    @PublishedApi
-    internal fun addImpl(node: T) {
-        assert { node.heap == null }
-        node.heap = this
-        val a = realloc()
-        val i = size++
-        a[i] = node
-        node.index = i
-        siftUpFrom(i)
-    }
-
-    private tailrec fun siftUpFrom(i: Int) {
-        if (i <= 0) return
-        val a = a!!
-        val j = (i - 1) / 2
-        if (a[j]!! <= a[i]!!) return
-        swap(i, j)
-        siftUpFrom(j)
-    }
-
-    private tailrec fun siftDownFrom(i: Int) {
-        var j = 2 * i + 1
-        if (j >= size) return
-        val a = a!!
-        if (j + 1 < size && a[j + 1]!! < a[j]!!) j++
-        if (a[i]!! <= a[j]!!) return
-        swap(i, j)
-        siftDownFrom(j)
-    }
-
-    @Suppress("UNCHECKED_CAST")
-    private fun realloc(): Array<T?> {
-        val a = this.a
-        return when {
-            a == null -> (arrayOfNulls<ThreadSafeHeapNode>(4) as Array<T?>).also { this.a = it }
-            size >= a.size -> a.copyOf(size * 2).also { this.a = it }
-            else -> a
-        }
-    }
-
-    private fun swap(i: Int, j: Int) {
-        val a = a!!
-        val ni = a[j]!!
-        val nj = a[i]!!
-        a[i] = ni
-        a[j] = nj
-        ni.index = i
-        nj.index = j
-    }
-}
+@Suppress("NOTHING_TO_INLINE")
+internal actual inline fun <T> clear(a: Array<T?>) = a.fill(null)
