+++ kotlinx-coroutines-core/native/src/CoroutineContext.kt
@@ -5,30 +5,25 @@
 package kotlinx.coroutines
 
 import kotlin.coroutines.*
-import kotlinx.coroutines.internal.*
 
 private fun takeEventLoop(): EventLoopImpl =
     ThreadLocalEventLoop.currentOrNull() as? EventLoopImpl ?:
         error("There is no event loop. Use runBlocking { ... } to start one.")
 
-internal object DefaultExecutor : CoroutineDispatcher(), Delay {
+internal actual object DefaultExecutor : CoroutineDispatcher(), Delay {
     override fun dispatch(context: CoroutineContext, block: Runnable) =
         takeEventLoop().dispatch(context, block)
-    override fun scheduleResumeAfterDelay(time: Long, continuation: CancellableContinuation<Unit>) =
-        takeEventLoop().scheduleResumeAfterDelay(time, continuation)
-    override fun invokeOnTimeout(time: Long, block: Runnable): DisposableHandle =
-        takeEventLoop().invokeOnTimeout(time, block)
+    override fun scheduleResumeAfterDelay(timeMillis: Long, continuation: CancellableContinuation<Unit>) =
+        takeEventLoop().scheduleResumeAfterDelay(timeMillis, continuation)
+    override fun invokeOnTimeout(timeMillis: Long, block: Runnable): DisposableHandle =
+        takeEventLoop().invokeOnTimeout(timeMillis, block)
 
-    fun enqueue(task: Runnable): Boolean {
+    actual fun enqueue(task: Runnable) {
         error("Cannot execute task because event loop was shut down")
     }
 
-    fun schedule(delayedTask: EventLoopImpl.DelayedTask) {
+    actual fun schedule(delayedTask: EventLoopImplBase.DelayedTask) {
         error("Cannot schedule task because event loop was shut down")
-    }
-
-    fun removeDelayedImpl(delayedTask: EventLoopImpl.DelayedTask) {
-        error("Cannot happen")
     }
 }
 
@@ -39,8 +34,8 @@ internal actual val DefaultDelay: Delay = DefaultExecutor
 
 public actual fun CoroutineScope.newCoroutineContext(context: CoroutineContext): CoroutineContext {
     val combined = coroutineContext + context
-    return if (combined !== kotlinx.coroutines.DefaultExecutor && combined[ContinuationInterceptor] == null)
-        combined + kotlinx.coroutines.DefaultExecutor else combined
+    return if (combined !== DefaultExecutor && combined[ContinuationInterceptor] == null)
+        combined + DefaultExecutor else combined
 }
 
 // No debugging facilities on native
