+++ kotlinx-coroutines-core/native/src/internal/Synchronized.kt
@@ -4,9 +4,17 @@
 
 package kotlinx.coroutines.internal
 
-@Suppress("ACTUAL_WITHOUT_EXPECT") // visibility
-internal actual typealias SynchronizedObject = Any
+import kotlinx.coroutines.*
 
-@PublishedApi
-internal actual inline fun <T> synchronized(lock: SynchronizedObject, block: () -> T): T =
+/**
+ * @suppress **This an internal API and should not be used from general code.**
+ */
+@InternalCoroutinesApi
+public actual typealias SynchronizedObject = Any
+
+/**
+ * @suppress **This an internal API and should not be used from general code.**
+ */
+@InternalCoroutinesApi
+public actual inline fun <T> synchronized(lock: SynchronizedObject, block: () -> T): T =
     block()
